Bot API 6.4 Update
December 30, 2022
Bot API 6.4

Added the field is_persistent to the class ReplyKeyboardMarkup, allowing to control when the keyboard is shown.
Added the parameter has_spoiler to the methods sendPhoto, sendVideo, and sendAnimation.
Added the field has_spoiler to the classes InputMediaPhoto, InputMediaVideo, and InputMediaAnimation.
Added the field has_media_spoiler to the class Message.
The parameters name and icon_custom_emoji_id of the method editForumTopic are now optional. If they are omitted, the existing values are kept.
Added the classes ForumTopicEdited, GeneralForumTopicHidden, GeneralForumTopicUnhidden, and WriteAccessAllowed and the fields forum_topic_edited, general_forum_topic_hidden, general_forum_topic_unhidden, and write_access_allowed to the class Message.
Added the methods editGeneralForumTopic, closeGeneralForumTopic, reopenGeneralForumTopic, hideGeneralForumTopic, unhideGeneralForumTopic for managing the General topic in forums.
Added the parameter message_thread_id to the method sendChatAction for sending chat actions to a specific message thread or a forum topic.
Added the field has_hidden_members to the class Chat. Note that the method getChatMember is only guaranteed to work if the bot is an administrator in the chat.
Added the field has_aggressive_anti_spam_enabled to the class Chat.
Added Web App events qrTextReceived and clipboardTextReceived.
Added the field platform to the class WebApp.
Added the methods showScanQrPopup, closeScanQrPopup, and readTextFromClipboard to the class WebApp.
Added the parameter options to the method openLink of the class WebApp.