Latest Updates Bot API 7.1, 7.2 and 7.3
May 6, 2024
Bot API 7.3

Added support for InlineKeyboardMarkup with url, login_url, and callback_game buttons for messages sent on behalf of a business account.
Added the field via_join_request to the class ChatMemberUpdated.
Added support for live locations that can be edited indefinitely, allowing 0x7FFFFFFF to be used as live_period.
Added the parameter live_period to the method editMessageLiveLocation.
Added the field question_entities to the class Poll.
Added the field text_entities to the class PollOption.
Added the parameters question_parse_mode and question_entities to the method sendPoll.
Added the class InputPollOption and changed the type of the parameter options in the method sendPoll to Array of InputPollOption.
Added the classes ChatBackground, BackgroundType, BackgroundFill and the field chat_background_set of type ChatBackground to the class Message, describing service messages about background changes.
Split out the class ChatFullInfo from the class Chat and changed the return type of the method getChat to ChatFullInfo.
Added the field max_reaction_count to the class ChatFullInfo.
Documented that .MP3 and .M4A files can be used as voice messages.
March 31, 2024
Bot API 7.2

Integration with Business Accounts

Added the class BusinessConnection and updates about the connection or disconnection of the bot to a business account, represented by the field business_connection in the class Update.
Added updates about new messages in a business account connected to the bot, represented by the field business_message in the class Update.
Added updates about message edits in a business account connected to the bot, represented by the field edited_business_message in the class Update.
Added updates about message deletion in a business account connected to the bot, represented by the class BusinessMessagesDeleted and the field deleted_business_messages in the class Update.
Added the method getBusinessConnection.
Working on Behalf of Business Accounts

Added the parameter business_connection_id to the methods sendMessage, sendPhoto, sendVideo, sendAnimation, sendAudio, sendDocument, sendSticker, sendVideoNote, sendVoice, sendLocation, sendVenue, sendContact, sendPoll, sendDice, sendGame, and sendMediaGroup.
Added the parameter business_connection_id to the method sendChatAction.
Added the field business_connection_id to the class Message.
Added the field sender_business_bot to the class Message.
Information about Business Accounts

Added the class BusinessIntro and the field business_intro to the class Chat.
Added the class BusinessLocation and the field business_location to the class Chat.
Added the classes BusinessOpeningHours and BusinessOpeningHoursInterval and the field business_opening_hours to the class Chat.
Mixed-Format Sticker Packs

Removed the fields is_animated and is_video from the class StickerSet.
Added the field format to the class InputSticker.
Removed the parameter sticker_format from the method createNewStickerSet.
Added the parameter format to the method setStickerSetThumbnail.
Increased the maximum number of stickers in any regular and mask sticker set to 120.
Allowed to upload WEBM stickers using sendSticker.
Request Chat Improvements

Added the fields request_name, request_username, and request_photo to the class KeyboardButtonRequestUsers.
Added the fields request_title, request_username, and request_photo to the class KeyboardButtonRequestChat.
Added the class SharedUser and replaced the field user_ids in the class UsersShared with the field users.
Added the fields title, username, and photo to the class ChatShared.
Other Changes

Added the field is_from_offline to the class Message.
Added the field can_connect_to_business to the class User.
Added the field personal_chat to the class Chat.
Added the method replaceStickerInSet,
Added the class Birthdate and the field birthdate to the class Chat.
Added the field BiometricManager to the class WebApp.
February 16, 2024
Bot API 7.1

Added support for the administrator rights can_post_stories, can_edit_stories, can_delete_stories in supergroups.
Added the class ChatBoostAdded and the field boost_added to the class Message for service messages about a user boosting a chat.
Added the field sender_boost_count to the class Message.
Added the field reply_to_story to the class Message.
Added the fields chat and id to the class Story.
Added the field unrestrict_boost_count to the class Chat.
Added the field custom_emoji_sticker_set_name to the class Chat.